activate data project

This function creates a new data project class instance in the global environment, updating the project's version

Usage

activate()

Value

data project class

Details

Data project properties

location - read only location path to the data project

name - read only name of the data project

version - read only version of the project

data_project - read only data project class exists in global environment

config - project set up to include the use of the config file to store path within the project

data - project set up to include data

databases - project set up to include the connection to databases and use of SQL queries

documentation - project set up to include package documentation

fme - project set up to include the use of FME workspaces and custom transformers

magrittr_pipe - project set up to use the magrittr's pipe, %>% in the package

package_version_control uses renv lock file to manage the package versions within the project

quality_assurance functions used for the validation and quality assurance of datasets

rtools - read only check that the correct version of rtools is available on the local machine

spreadsheets - project set up to include the use of spreadsheets

tasks - project uses tasks quarto document to list tasks for the project

typst_documents - project set up to include typst documents

unit_tests - project set up to include function unit tests using the testthat package

version_control - project set up to use Git as version control system

vignettes - project set up to include vignette quarto documents

Data project generic functions and methods

access_version_control - create and store a GitHub Personal Access token in the local credential store

add_data_script - includes the use of data within the project, adding a data-raw script for the dataset(s)

add_function_script - adds a functions script in the R folder, including suggested Roxygen comments

add_lookup_script - adds a lookup-raw script to the data-raw folder

add_quarto_vignette - adds a blank Quarto vignette to the vignettes folder

add_typst_document - adds a blank Typst document to the typst_docs folder

add_unit_test - setups the data project to use unit testing, adds a unit test file to the tests folder

document_all_functions_and_lookups - converts all the functions and lookup documentation into corresponding html files for documentation

create_new_package - opens a new package form in RStudio to run and fill in to create a new R package within the data project

document_data_project - updates the data project's documentation

document_home - renders the home page into corresponding markdown document for documentation

document_products - converts products document into corresponding html file for documentation

document_all_functions_and_lookups - converts all the functions and lookups markdown documents into corresponding html files for documentation

document_all_quarto_articles - converts all quarto articles to corresponding html files for documentation

document_single_quarto_article - converts single quarto article to corresponding html file for documentation

document_tasks - converts tasks quarto article to corresponding html file for documentation

include_config - included the use of the config file within the project

include_databases - includes the use of connecting to databases and SQL queries within the project

include_documentation - includes documentation in the project

include_fme - includes the use of FME workspaces and custom transformers within the project

include_magrittr_pipe - includes the use of the magrittr pipe in the project

include_quality_assurance - adds functions used for the validation and quality assurance of datasets

include_spreadsheets - includes the use of spreadsheets within the project

include_version_control - creates a Git and GitHub repository for the project

increment_major_version - increments major level of data project version, with message to update lifecycle badge

increment_minor_version - increments minor level of data project version, with message to update lifecycle badge

open_config - opens the config yaml and .Renviron files in RStudio for editing

open_description - opens the DESCRIPTION document in RStudio for editing

open_home - opens the README Rmarkdown document in RStudio for editing

open_pkgdown - opens the _pkgdown document in RStudio for editing

open_products - opens the NEWS document in RStudio for editing

open_tasks - opens the tasks document in RStudio for editing

sync_version_control - stages and commits data project files to the git repository