add data script

This generic function does the necessary setup to include data within the project by running usethis use_data_raw function (https://usethis.r-lib.org/reference/use_data.html) to add a data-raw script for the dataset and creating a data folder

Usage

add_data_script(x, ...)

Arguments

name

script_title character title for the data-raw script

Details

Data objects are saved within the data folder by running usethis use_data function

ROxygen comments added to function script in R folder

@title TITLE

@description

@details

  <R CODE CHUNK>
    DATASET_NAME |>
       dplyr::select(FIELDNAMES) |>
       gt::gt_preview(top_n = nrow(DATASET_NAME) |>
       gt::tab_style(style = gt::cell_text(style = "italic"),
                     locations = gt::cells_body(columns = name)) |>
       gt::tab_options(table.align='left')
  <R CODE CHUNK>

  @format A tibble with  <backslash> r nrow(DATASET_NAME) <backslash> rows and
          <backslash> r ncol(DATASET_NAME) <backslash> variables
  \describe{
    \item{FIELDNAME}{DESCRIPTION}
  }

"DATASET NAME"