deactivate data project

This function updates the data project prior to copying it and removing the data project class instance from the global environment

creating or updating a renv lock file to store data project's package dependencies versions

refreshing the list of package dependencies in package's description file

Usage

deactivate()

Details

The data project can be unzipped and copied onto a different machine, using the renv package to manage the data project's packages

renv::activate()

renv::restore(prompt = FALSE))