HellingerDist(distrEx)	R Documentation

Generic function for the computation of the Hellinger distance of two distributions

Description

Generic function for the computation of the Hellinger distance d_h of two distributions P and Q which may be defined for an arbitrary sample space (Omega, A). The Hellinger distance is defined as

0.5 int |sqrt{dP}-sqrt{dQ}|^2

where sqrt{dP}, respectively sqrt{dQ} denotes the square root of the densities.

Usage


HellingerDist(e1, e2, ...)
## S4 method for signature 'AbscontDistribution,
##   AbscontDistribution':
HellingerDist(e1,e2)
## S4 method for signature 'AbscontDistribution,
##   DiscreteDistribution':
HellingerDist(e1,e2)
## S4 method for signature 'DiscreteDistribution,
##   AbscontDistribution':
HellingerDist(e1,e2)
## S4 method for signature 'DiscreteDistribution,
##   DiscreteDistribution':
HellingerDist(e1,e2)

Arguments

e1 	 object of class "Distribution"

e2 	 object of class "Distribution"

... 	 further arguments to be used in particular methods (not in package distrEx)

Value

A list containing the following components:

e1 	 object of class "Distribution"; distribution 1

e2 	 object of class "Distribution"; distribution 2

Hellinger.distance 	 Hellinger distance of e1 and e2

Methods

e1 = "AbscontDistribution", e2 = "AbscontDistribution":

Hellinger distance of two absolutely continuous univariate distributions which is computed using distrExintegrate.

e1 = "AbscontDistribution", e2 = "DiscreteDistribution":

Hellinger distance of absolutely continuous and discrete univariate distributions (are mutually singular; i.e., have distance =1).

e1 = "DiscreteDistribution", e2 = "DiscreteDistribution":

Hellinger distance of two discrete univariate distributions which is computed using support and sum.

e1 = "DiscreteDistribution", e2 = "AbscontDistribution":

Hellinger distance of discrete and absolutely continuous univariate distributions (are mutually singular; i.e., have distance =1).

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

See Also

distrExIntegrate, ContaminationSize, TotalVarDist, KolmogorovDist, Distribution-class

Examples


HellingerDist(Norm(), Gumbel())
HellingerDist(Norm(), Td(10))
HellingerDist(Norm(mean = 50, sd = sqrt(25)), Binom(size = 100)) # mutually singular
HellingerDist(Pois(10), Binom(size = 20)) 

[Package distrEx version 2.0 Index]